strict-dynamic
